feat: added a second footer type that can be used when type is specified as "menu"
