+++ playground/pages/index.vue
@@ -1,15 +1,23 @@
 <script setup lang="ts">
-const route = useRoute()
+const links = reactive([{
+  linkName: 'Home',
+  linkAddress: '/',
+}])
 </script>
 
 <template>
   <div>
-    <h1>Nuxt Routing set up successfully!</h1>
-    <p>Current route: {{ route.path }}</p>
-    <a
-      href="https://nuxt.com/docs/getting-started/routing"
-      target="_blank"
-    >Learn more about Nuxt Routing</a>
-    <s-SimpleCard title="Sample Card">Hello</s-SimpleCard>
+    <s-MainHeader
+      site-logo="/logo.webp"
+      font-color="text-sky-900"
+      :page-links="links"
+    />
+    <s-SimpleCard>
+      Hello
+    </s-SimpleCard>
+    <s-SimpleFooter
+      copyright="Brendan"
+      type="menu"
+    />
   </div>
 </template>
