+++ src/runtime/components/Types.ts
@@ -1,0 +1,4 @@
+export interface MenuObject {
+  linkName: string
+  linkAddress: string
+}
